# stewardess_openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stewardess MCP Service",
    "description": "Local source-code repository tools for AI agents. Provides file reading, search, editing, git, and command execution.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://office:55702"
    }
  ],
  "paths": {
...
    "/api/command/build": {
      "post": {
        "operationId": "Command_RunBuild",
        "summary": "Runs the configured build command (e.g. dotnet build or msbuild).",
        "description": "Runs the configured build command (e.g. dotnet build or msbuild).",
        "tags": [
          "Command"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBuildRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/command/test": {
      "post": {
        "operationId": "Command_RunTests",
        "summary": "Runs the configured test command (e.g. dotnet test).",
        "description": "Runs the configured test command (e.g. dotnet test).",
        "tags": [
          "Command"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunTestsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/command/run": {
      "post": {
        "operationId": "Command_RunCustomCommand",
        "summary": "Runs a custom command that must appear in the AllowedCommands whitelist.",
        "description": "Runs a custom command that must appear in the AllowedCommands whitelist.",
        "tags": [
          "Command"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCustomCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RunBuildRequest": {
        "description": "Request to run the repository's build command.",
        "type": "object",
        "properties": {
          "workingDirectory": {
            "description": "Working directory relative to repository root.  Empty = root.\r\nTypically the directory containing a .sln or .csproj file.",
            "type": "string"
          },
          "buildCommand": {
            "description": "Build tool to use: \"dotnet build\", \"msbuild\", or another allowed command.\r\nDefaults to \"dotnet build\".",
            "type": "string"
          },
          "arguments": {
            "description": "Additional CLI arguments appended to the build command.",
            "type": "string"
          },
          "configuration": {
            "description": "Target configuration: \"Debug\" (default) or \"Release\".",
            "type": "string"
          },
          "timeoutSeconds": {
            "format": "int32",
            "description": "Override the server's timeout (seconds); capped at MaxCommandExecutionSeconds.",
            "type": "integer"
          }
        }
      },
...
      "RunTestsRequest": {
        "description": "Request to run the repository's test suite.",
        "type": "object",
        "properties": {
          "workingDirectory": {
            "description": "Working directory relative to repository root. Empty = root.",
            "type": "string"
          },
          "testCommand": {
            "description": "Test command: \"dotnet test\" (default) or another allowed command.",
            "type": "string"
          },
          "arguments": {
            "description": "Extra arguments appended to the test command.",
            "type": "string"
          },
          "filter": {
            "description": "Filter expression passed to the test runner (e.g. dotnet test --filter).",
            "type": "string"
          },
          "configuration": {
            "description": "Build configuration to use, e.g. \"Debug\" or \"Release\".",
            "type": "string"
          },
          "timeoutSeconds": {
            "format": "int32",
            "description": "Optional override for the maximum allowed execution time in seconds.",
            "type": "integer"
          }
        }
      },
      "RunCustomCommandRequest": {
        "description": "Request to execute a custom command from the server's AllowedCommands whitelist.",
        "type": "object",
        "properties": {
          "command": {
            "description": "The full command line to execute.  Must start with an entry in AllowedCommands.",
            "type": "string"
          },
          "workingDirectory": {
            "description": "Working directory relative to repository root.  Empty = root.",
            "type": "string"
          },
          "timeoutSeconds": {
            "format": "int32",
            "description": "Optional override for the maximum allowed execution time in seconds.",
            "type": "integer"
          }
        }
      },
...
    }
  }
}